global: add testlib module for reusability
* Adds celery_monitor module to testlib.
* Adds tasks module to testlib.
* Adds custom scrapyd runner for coverage.
* Adds `fixtures` module instead of `__init__.py` for useful test utils.
* Re-factored `fixtures.py`.

Closes #100
