+++ hepcrawl/testlib/fixtures.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of hepcrawl.
-# Copyright (C) 2015 CERN.
+# Copyright (C) 2015, 2016, 2017 CERN.
 #
 # hepcrawl is a free software; you can redistribute it and/or modify it
 # under the terms of the Revised BSD License; see LICENSE file for
 # more details.
+
+from __future__ import absolute_import, division, print_function
 
 import os
 
@@ -24,12 +26,10 @@ def fake_response_from_file(file_name, url='http://www.example.com', response_ty
 
     :returns: A scrapy HTTP response which can be used for unittesting.
     """
-    meta = {}
     request = Request(url=url)
 
     if not file_name[0] == '/':
-        responses_dir = os.path.dirname(os.path.realpath(__file__))
-        file_path = os.path.join(responses_dir, file_name)
+        file_path = get_responses_path(file_name)
     else:
         file_path = file_name
 
@@ -47,7 +47,6 @@ def fake_response_from_file(file_name, url='http://www.example.com', response_ty
 
 def fake_response_from_string(text, url='http://www.example.com', response_type=TextResponse):
     """Fake Scrapy response from a string."""
-    meta = {}
     request = Request(url=url)
     response = response_type(
         url=url,
@@ -65,6 +64,31 @@ def get_node(spider, tag, response=None, text=None, rtype="xml"):
         selector = Selector(response, type=rtype)
     elif text:
         selector = Selector(text=text, type=rtype)
+
     spider._register_namespaces(selector)
     node = selector.xpath(tag)
     return node
+
+
+def get_responses_path(*path_chunks):
+    """
+    :param path_chunks: Optional extra path element to suffix the responses directory with.
+    :return: The absolute path to the responses and if path_chuncks provided the absolute
+     path to path chunks.
+
+    :Example:
+
+        >>> get_responses_path()
+        '/home/myuser/hepcrawl/tests/responses'
+
+        >>> get_responses_path('one', 'two')
+        '/home/myuser/hepcrawl/tests/responses/one/two'
+    """
+    project_root_dir = os.path.abspath(
+        os.path.join(
+            os.path.dirname(os.path.abspath(__file__)),
+            '..',
+            '..',
+        )
+    )
+    return os.path.join(project_root_dir, 'tests', 'unit', 'responses', *path_chunks)
