+++ hepcrawl/testlib/scrapyd_coverage_runner.py
