+++ tests/docker_entrypoint.sh
