--- tests/functional/WSP/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# This file is part of hepcrawl.
-# Copyright (C) 2015, 2016, 2017 CERN.
-#
-# hepcrawl is a free software; you can redistribute it and/or modify it
-# under the terms of the Revised BSD License; see LICENSE file for
-# more details.
