+++ tests/functional/scrapyd_coverage_runner.conf
@@ -9,4 +9,4 @@
 
 
 [scrapyd]
-runner = tests.functional.scrapyd_coverage_runner
+runner = hepcrawl.testlib.scrapyd_coverage_runner
