+++ tests/unit/test_aps.py
@@ -12,7 +12,7 @@ from __future__ import absolute_import, print_function, unicode_literals
 import pytest
 
 from hepcrawl.spiders import aps_spider
-from .responses import fake_response_from_file
+from hepcrawl.testlib.fixtures import fake_response_from_file
 
 
 @pytest.fixture
@@ -29,6 +29,7 @@ def results():
             )
         )
     )
+
     assert records
     return records
