+++ tests/unit/test_arxiv_single.py
@@ -17,7 +17,7 @@ from inspire_schemas.api import validate
 
 from hepcrawl.pipelines import InspireCeleryPushPipeline
 from hepcrawl.spiders import arxiv_spider
-from .responses import fake_response_from_file
+from hepcrawl.testlib.fixtures import fake_response_from_file
 
 
 @pytest.fixture
@@ -27,12 +27,15 @@ def results():
 
     crawler = Crawler(spidercls=arxiv_spider.ArxivSpider)
     spider = arxiv_spider.ArxivSpider.from_crawler(crawler)
-    records = list(spider.parse(
-        fake_response_from_file(
-            'arxiv/sample_arxiv_record0.xml',
-            response_type=TextResponse
+    records = list(
+        spider.parse(
+            fake_response_from_file(
+                'arxiv/sample_arxiv_record0.xml',
+                response_type=TextResponse,
+            )
         )
-    ))
+    )
+
     assert records
     pipeline = InspireCeleryPushPipeline()
     pipeline.open_spider(spider)
