+++ tests/unit/test_brown.py
@@ -19,7 +19,7 @@ import hepcrawl
 
 from hepcrawl.spiders import brown_spider
 
-from .responses import (
+from hepcrawl.testlib.fixtures import (
     fake_response_from_file,
     fake_response_from_string,
 )
