+++ tests/unit/test_edp.py
@@ -21,7 +21,7 @@ from hepcrawl.items import HEPRecord
 
 from scrapy.http import HtmlResponse, Request
 
-from .responses import (
+from hepcrawl.testlib.fixtures import (
     fake_response_from_file,
     fake_response_from_string,
     get_node,
