+++ tests/unit/test_elsevier.py
@@ -15,7 +15,7 @@ import pytest
 
 from hepcrawl.spiders import elsevier_spider
 
-from .responses import (
+from hepcrawl.testlib.fixtures import (
     fake_response_from_file,
     fake_response_from_string,
     get_node,
