+++ tests/unit/test_magic.py
@@ -17,7 +17,7 @@ import hepcrawl
 
 from hepcrawl.spiders import magic_spider
 
-from .responses import (
+from hepcrawl.testlib.fixtures import (
     fake_response_from_file,
     fake_response_from_string,
     get_node,
