+++ tests/unit/test_mit.py
@@ -13,7 +13,7 @@ import pytest
 
 from hepcrawl.spiders import mit_spider
 
-from .responses import (
+from hepcrawl.testlib.fixtures import (
     fake_response_from_file,
     fake_response_from_string,
     get_node,
