+++ tests/unit/test_phil.py
@@ -15,7 +15,7 @@ import pytest
 
 from hepcrawl.spiders import phil_spider
 
-from .responses import fake_response_from_file
+from hepcrawl.testlib.fixtures import fake_response_from_file
 
 
 @pytest.fixture
