+++ tests/unit/test_world_scientific.py
@@ -18,7 +18,7 @@ from scrapy.http import TextResponse
 from hepcrawl.pipelines import InspireCeleryPushPipeline
 from hepcrawl.spiders import wsp_spider
 
-from .responses import fake_response_from_file
+from hepcrawl.testlib.fixtures import fake_response_from_file
 
 
 def create_spider():
